+++ packages/headless/src/features/query/query-slice.ts
@@ -4,6 +4,7 @@ import {selectQuerySuggestion} from '../query-suggest/query-suggest-actions';
 import {change} from '../history/history-actions';
 import {applyDidYouMeanCorrection} from '../did-you-mean/did-you-mean-actions';
 import {getQueryInitialState} from './query-state';
+import {restoreSearchParameters} from '../search-parameters/search-parameter-actions';
 
 export const queryReducer = createReducer(getQueryInitialState(), (builder) =>
   builder
@@ -14,7 +15,10 @@ export const queryReducer = createReducer(getQueryInitialState(), (builder) =>
     .addCase(selectQuerySuggestion, (state, action) => {
       state.q = action.payload.expression;
     })
-    .addCase(change.fulfilled, (state, action) => {
-      state.q = action.payload.query.q;
+    .addCase(change.fulfilled, (_, action) => action.payload.query)
+    .addCase(restoreSearchParameters, (state, action) => {
+      state.q = action.payload.q ?? state.q;
+      state.enableQuerySyntax =
+        action.payload.enableQuerySyntax ?? state.enableQuerySyntax;
     })
 );